move TextFormField to separate file
